优化----网关心跳优化兼容并发发送心跳的情况

+++ rundo-gb28181-server/src/main/java/com/runjian/gb28181/bean/Device.java
@@ -104,7 +104,7 @@ public class Device {
 	 * 注册有效期
 	 */
 	@Schema(description = "心跳有效期")
-	private long keepaliveIntervalTime;
+	private Integer keepaliveIntervalTime;
 
 	/**
 	 * 注册有效期
